requirements.txt
certifi==2021.5.30
chardet==4.0.0
colorful==0.5.4
idna==2.10
Jinja2==3.0.1
MarkupSafe==2.0.1
prettyprinter==0.18.0
Pygments==2.9.0
PyYAML==5.4.1
requests==2.25.1
urllib3==1.26.5
